+++ hasher-matcher-actioner/hma-lite/hmalite/app.py
@@ -8,6 +8,7 @@ import shutil
 from flask import Flask, render_template, request, jsonify, send_from_directory
 from hmalite.config import HmaLiteDevConfig, HmaLiteProdConfig
 from hmalite.matcher import matcher_api
+from hmalite.index import reset_index, get_local_index
 from threatexchange.hashing.pdq_hasher import pdq_from_file
 from threatexchange.signal_type import pdq_index
 
@@ -28,27 +29,6 @@ if better_config.DEBUG:
     app.logger.info("Config Values:")
     for name in config_cls._fields:
         app.logger.info("%s = %s", name, app.config[name])
-
-
-####################### INDEX HACKS #######################
-# This doesn't really behave the way you might think -
-# instances of flask are destroyed and created all the time/thread
-# local-d, so you'll get a bunch of copies of this(?)
-# Someday we'll have to figure out the correct way to stage it
-_INDEX = None
-
-def get_local_index():
-    global _INDEX
-    if not _INDEX:
-        filepath = better_config.local_index_file
-        with open(filepath, "rb") as f:
-            index = pdq_index.PDQIndex.deserialize(f.read())
-        _INDEX = index
-    return _INDEX
-
-def reset_index(index):
-    global _INDEX
-    _INDEX = index
 
 
 #################### VARIOUS ENDPOINTS ####################
@@ -115,9 +95,7 @@ def create_index(filepath):
         if sniffer.has_header(sample):
             reader = csv.DictReader(csvfile)
         else:
-            reader = csv.DictReader(
-                csvfile, fieldnames=["hash"], restkey="meta"
-            )
+            reader = csv.DictReader(csvfile, fieldnames=["hash"], restkey="meta")
 
         # Try and guess what column has the hash
         key = reader.fieldnames[0]
